[brownfield][js] reorganize src structure to better separate module specific code
