+++ packages/expo-brownfield/package.json
@@ -63,6 +63,7 @@
     "expo-module-scripts": "^55.0.2"
   },
   "peerDependencies": {
-    "expo": "*"
+    "expo": "*",
+    "react": "*"
   }
 }
