+++ packages/expo-brownfield/src/ExpoBrownfieldStateModule.types.ts
@@ -1,0 +1,6 @@
+import type { NativeModule } from 'expo';
+
+export declare class ExpoBrownfieldStateModuleSpec extends NativeModule {
+  getSharedState(key: string): any;
+  deleteSharedState(key: string): void;
+}
